Extract fake controller to use it in other tests
